requirements.txt
Flask==2.3.3
requests==2.31.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.97.0
python-dateutil==2.8.2
Werkzeug==2.3.7
gunicorn==21.2.0

# Database (SQLite only)
Flask-SQLAlchemy==3.0.5

# Security and validation
cryptography==41.0.3
itsdangerous==2.1.2

# Utils
click==8.1.7
python-dotenv==1.0.0
